requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.3
Django==1.11.6
django-bootstrap-toolkit==2.15.0
django-environ==0.4.5
django-registration-redux==1.4
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
uritemplate==3.0.0
urllib3==1.25.3
